requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.9

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── AI / ML ──────────────────────────────────────────────────────────────────
google-generativeai==0.7.2
faiss-cpu==1.8.0
numpy==1.26.4

# ── Async I/O ────────────────────────────────────────────────────────────────
aiofiles==24.1.0
httpx==0.27.2

# ── Utilities ────────────────────────────────────────────────────────────────
structlog==24.4.0
python-dotenv==1.0.1
tenacity==9.0.0

# ── Dev / test ────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2          # for TestClient
